Refactor logistics controller and module
