+++ scms_backend/src/logistics/logistics.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { LogisticsController } from './logistics.controller';
-import { LogisticsService } from './logistics.service';
+import { LogisticsController } from './controllers/logistics.controller';
+import { LogisticsService } from './services/logistics.service';
 import { DriverService } from './services/driver.service';
 import { TypeOrmModule } from '@nestjs/typeorm/dist/typeorm.module';
 import { Driver } from './entities/driver.entity';
