+++ scms_backend/src/logistics/services/logistics.service.ts
@@ -3,7 +3,6 @@ import { Injectable } from '@nestjs/common';
 @Injectable()
 export class LogisticsService {
   welcome() {
-    // Logic to retrieve the list of available transportation options
-    return 'Hello LM';
+    return 'Welcome to the Logistics API!';
   }
 }
